voice-gateway/requirements.txt
# Voice Gateway Dependencies
# v2.0.0

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# HTTP Client
httpx==0.28.1

# AWS SDK
boto3==1.35.87

# AWS Transcribe Streaming
amazon-transcribe==0.6.4

# Environment
python-dotenv==1.0.1

# Data Validation (FastAPI dependency)
pydantic>=2.0.0

# Async Support
aiofiles>=23.0.0
